+++ pkg/patterns/messaging/message.go
