+++ pkg/patterns/messaging/publisher/errors.go
