+++ pkg/platform/gcp/pubsub/client.go
@@ -3,10 +3,10 @@ package pubsub
 
 import (
 	"context"
+	"github.com/marcodd23/go-micro-core/pkg/patterns/messaging"
+	"github.com/marcodd23/go-micro-core/pkg/patterns/messaging/publisher"
 
 	"cloud.google.com/go/pubsub"
-	"github.com/marcodd23/go-micro-core/pkg/messaging"
-	"github.com/marcodd23/go-micro-core/pkg/messaging/publisher"
 )
 
 // ============================================
