+++ pkg/platform/gcp/pubsub/publisher.go
@@ -2,10 +2,10 @@ package pubsub
 
 import (
 	"context"
+	publisher2 "github.com/marcodd23/go-micro-core/pkg/patterns/messaging/publisher"
 	"time"
 
 	"cloud.google.com/go/pubsub"
-	"github.com/marcodd23/go-micro-core/pkg/messaging/publisher"
 	"google.golang.org/api/option"
 )
 
@@ -15,25 +15,25 @@ func NewPubSubBufferedPublisherFactory(
 	projectID string,
 	batchSize int32,
 	flushDelayThreshold int32,
-	opts ...option.ClientOption) (publisher.BufferedPublisher, error) {
+	opts ...option.ClientOption) (publisher2.BufferedPublisher, error) {
 	client, err := pubsub.NewClient(ctx, projectID, opts...)
 	if err != nil {
-		return nil, publisher.NewMessagingErrorCode(publisher.ErrorInitializingPubsubClient, err)
+		return nil, publisher2.NewMessagingErrorCode(publisher2.ErrorInitializingPubsubClient, err)
 	}
 
-	publishConfig := publisher.TopicPublishConfig{
+	publishConfig := publisher2.TopicPublishConfig{
 		BatchSize: batchSize,
 	}
 
 	if flushDelayThreshold <= 0 {
-		publishConfig.FlushDelayThreshold = publisher.DefaultFlushDelayThreshold
+		publishConfig.FlushDelayThreshold = publisher2.DefaultFlushDelayThreshold
 	} else {
 		publishConfig.FlushDelayThreshold = time.Duration(flushDelayThreshold) * time.Millisecond
 	}
 
 	pubSubClient := &pubSubClient{client: client}
 
-	return publisher.NewBufferedPublisher(pubSubClient, publishConfig)
+	return publisher2.NewBufferedPublisher(pubSubClient, publishConfig)
 }
 
 // NewBufferedPublisherWithRetryFactory - factory that create a cloud_pubsub client and then initialize a publisher.BufferedPublisherWithRetry.
@@ -44,39 +44,39 @@ func NewBufferedPublisherWithRetryFactory(
 	flushDelayThresholdMillis int32,
 	maxRetryCount int16,
 	initialRetryIntervalMillis int32,
-	opts ...option.ClientOption) (publisher.BufferedPublisherWithRetry, error) {
+	opts ...option.ClientOption) (publisher2.BufferedPublisherWithRetry, error) {
 	client, err := pubsub.NewClient(ctx, projectID, opts...)
 	if err != nil {
-		return nil, publisher.NewMessagingErrorCode(publisher.ErrorInitializingPubsubClient, err)
+		return nil, publisher2.NewMessagingErrorCode(publisher2.ErrorInitializingPubsubClient, err)
 	}
 
-	publishConfig := publisher.TopicPublishConfig{}
+	publishConfig := publisher2.TopicPublishConfig{}
 
 	if batchSize <= 0 {
-		publishConfig.BatchSize = publisher.DefaultBatchSize
+		publishConfig.BatchSize = publisher2.DefaultBatchSize
 	} else {
 		publishConfig.BatchSize = batchSize
 	}
 
 	if flushDelayThresholdMillis <= 0 {
-		publishConfig.FlushDelayThreshold = publisher.DefaultFlushDelayThreshold
+		publishConfig.FlushDelayThreshold = publisher2.DefaultFlushDelayThreshold
 	} else {
 		publishConfig.FlushDelayThreshold = time.Duration(flushDelayThresholdMillis) * time.Millisecond
 	}
 
 	if initialRetryIntervalMillis <= 0 {
-		publishConfig.InitialRetryInterval = publisher.DefaultInitialRetryInterval
+		publishConfig.InitialRetryInterval = publisher2.DefaultInitialRetryInterval
 	} else {
 		publishConfig.InitialRetryInterval = time.Duration(initialRetryIntervalMillis) * time.Millisecond
 	}
 
 	if maxRetryCount <= 0 {
-		publishConfig.MaxRetryCount = publisher.DefaultMaxRetryCount
+		publishConfig.MaxRetryCount = publisher2.DefaultMaxRetryCount
 	} else {
 		publishConfig.MaxRetryCount = maxRetryCount
 	}
 
 	pubSubClient := &pubSubClient{client: client}
 
-	return publisher.NewBufferedPublisherWithRetry(ctx, pubSubClient, publishConfig)
+	return publisher2.NewBufferedPublisherWithRetry(ctx, pubSubClient, publishConfig)
 }
